+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Inventory Camera (OpenMW)
 
+## 1.1
+
+- Expanded FAQ section
+- Added "Disable Camera Change While ___ is Held" setting (?)
+- Added "Disable Camera Change While in Combat" setting (?)
+- Fixed player rotating from the camera when using items from the inventory (?)
+
 ## 1.0
 
 Initial release